Move to ACAP 4 Native SDK and open62541 1.4.4
This patchset moves to the ACAP 4 Native SDK. This means we drop
support for (really) old devices. If you need to build with the ACAP 3
SDK for old devices, please use the previous commit as your base since
that will work fine for that.

With the move to ACAP 4 Native SDK, we can drop the support for this
repo to be built with locally installed SDKs (that nobody uses these
days anyway) and then move the build of open62541 from Makefile into
the Dockerfile instead.

The patchset also bumps open62541 to version 1.4.4.

Change-Id: Idba0573102fe52df0d6239a300185c36f4c6bc94
Signed-off-by: Joakim Roubert <joakimr@axis.com>